+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to the "lua-tags" extension will be documented in this file.
 1. fix require path auto completion show duplicate path
 2. fix permission deny on linux
 3. allow jump to local unreachable symbol definition
-4. do't do completion when definte a function
+4. do't do completion or signature when definte a function
 
 ## [1.0.2] 2019-12-18
 1. set luacheck timeout to 15s